fix(supervisor): supervisor v2.1.1
新增修改密码接口；修改了Supervisor表中gender类型，改为int;

+++ src/main/java/com/example/heart_field/entity/Supervisor.java
@@ -17,7 +17,7 @@ public class Supervisor {
     private LocalDateTime updateTime;
     private String name;
     private Integer age;
-    private byte gender;
+    private int gender;
     private String password;
     private String avatar;
     private Integer maxConcurrent;
